+++ RequestInfo/m+mRequestInfoMain.cpp
@@ -322,7 +322,7 @@ static void setUpAndGo(const YarpString &  channelName,
     }
     if (Utilities::GetServiceNamesFromCriteria(channelNameRequest, services))
     {
-        int matchesCount = services.size();
+        size_t matchesCount = services.size();
 
         if (matchesCount)
         {
